[Fix]: new flow in dialog for delete task

+++ uca_walkmate/lib/presentation/widgets/menu/subject_delete_popup.dart
@@ -8,23 +8,43 @@ class SubjectDeletePopup extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return IconButton(
+    return PopupMenuButton<String>(
       icon: const Icon(Icons.more_vert),
-      onPressed: () {
-        showMenu(
-          context: context,
-          position: const RelativeRect.fromLTRB(25.0, 25.0, 0.0, 0.0),
-          items: [
-            const PopupMenuItem<String>(
-              child: Text('Eliminar materia'),
-            ),
-          ],
-        ).then((String? item) {
-          
-            ref.read(subjectProvider.notifier).deleteSubjectById(id);
-          
-        });
+      onSelected: (String result) {
+        if (result == 'delete') {
+          showDialog(
+            context: context,
+            builder: (BuildContext context) {
+              return AlertDialog(
+                title: const Text('Confirmación'),
+                content: const Text(
+                    '¿Estás seguro que deseas eliminar esta materia?'),
+                actions: [
+                  TextButton(
+                    onPressed: () {
+                      Navigator.of(context).pop(false);
+                    },
+                    child: const Text('Cancelar'),
+                  ),
+                  TextButton(
+                    onPressed: () {
+                      ref.read(subjectProvider.notifier).deleteSubjectById(id);
+                      Navigator.of(context).pop(true);
+                    },
+                    child: const Text('Eliminar'),
+                  ),
+                ],
+              );
+            },
+          );
+        }
       },
+      itemBuilder: (BuildContext context) => [
+        const PopupMenuItem<String>(
+          value: 'delete',
+          child: Text('Eliminar materia'),
+        ),
+      ],
     );
   }
 }
